requirements.txt
praw
edge-tts
torch --index-url https://download.pytorch.org/whl/cu118
python-dotenv
markdown
Pillow
moviepy